pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "common-assessment-tool-CommitMaster"
version = "0.0.1"
authors = [
  { name = "Zella", email = "zellayu1212@gmail.com" },
  { name = "Xinyi", email = "xysh3n@gmail.com" },
  { name = "ZeYuan", email = "997852533@qq.com" },
  { name = "Bethany", email = "bethanyzhangdan@gmail.com" }
]
description = "A FastAPI backend application with ML model switching support"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Framework :: FastAPI",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
    "Jinja2==3.1.2",
    "Mako==1.2.4",
    "MarkupSafe==2.1.3",
    "PyYAML==6.0.1",
    "Pygments==2.16.1",
    "QtPy==2.3.1",
    "SQLAlchemy==2.0.21",
    "Send2Trash==1.8.0",
    "aiofiles==23.2.1",
    "alembic==1.12.0",
    "annotated-types==0.6.0",
    "anyio==3.7.1",
    "appnope==0.1.3",
    "argon2-cffi-bindings==21.2.0",
    "argon2-cffi==21.3.0",
    "arrow==1.2.3",
    "astroid==3.0.1",
    "asttokens==2.4.0",
    "attrs==22.1.0",
    "backcall==0.2.0",
    "bcrypt==4.0.1",
    "beautifulsoup4==4.12.2",
    "black==23.10.0",
    "bleach==6.0.0",
    "branca==0.6.0",
    "certifi==2023.7.22",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "comm==0.1.3",
    "cryptography==41.0.4",
    "debugpy==1.6.7",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "dill==0.3.7",
    "ecdsa==0.18.0",
    "exceptiongroup==1.0.4",
    "executing==2.0.0",
    "fastapi==0.103.2",
    "fastjsonschema==2.16.3",
    "folium==0.14.0",
    "fqdn==1.5.1",
    "h11==0.14.0",
    "httptools==0.6.0",
    "httpx==0.24.1",
    "idna==3.4",
    "iniconfig==1.1.1",
    "ipykernel==6.22.0",
    "ipython-genutils==0.2.0",
    "ipython==8.16.1",
    "ipywidgets==8.0.6",
    "isoduration==20.11.0",
    "isort==5.12.0",
    "jedi==0.19.1",
    "joblib==1.4.0",
    "jsonpointer==2.3",
    "jsonschema==4.17.3",
    "jupyter-console==6.6.3",
    "jupyter-events==0.6.3",
    "jupyter==1.0.0",
    "jupyter_client==8.2.0",
    "jupyter_core==5.3.0",
    "jupyter_server==2.5.0",
    "jupyter_server_terminals==0.4.4",
    "jupyterlab-pygments==0.2.2",
    "jupyterlab-widgets==3.0.7",
    "lxml==4.9.3",
    "matplotlib-inline==0.1.6",
    "mccabe==0.7.0",
    "mistune==2.0.5",
    "mypy-extensions==1.0.0",
    "nbclassic==0.5.5",
    "nbclient==0.7.3",
    "nbconvert==7.3.1",
    "nbformat==5.8.0",
    "nest-asyncio==1.5.6",
    "notebook==6.5.4",
    "notebook_shim==0.2.2",
    "numpy==1.24.2",
    "packaging==23.2",
    "pandas==2.0.0",
    "pandocfilters==1.5.0",
    "parso==0.8.3",
    "passlib==1.7.4",
    "pathspec==0.11.2",
    "pexpect==4.8.0",
    "pickleshare==0.7.5",
    "platformdirs==3.11.0",
    "pluggy==1.0.0",
    "pre-commit",
    "prometheus-client==0.16.0",
    "prompt-toolkit==3.0.39",
    "psutil==5.9.5",
    "psycopg2-binary==2.9.9",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.2",
    "pyasn1==0.5.0",
    "pycodestyle==2.10.0",
    "pycparser==2.21",
    "pydantic-settings==2.0.3",
    "pydantic==2.4.2",
    "pydantic_core==2.10.1",
    "pylint==3.0.1",
    "pyrsistent==0.19.3",
    "pytest==7.2.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.6",
    "pytz==2023.3",
    "pyzmq==25.0.2",
    "qtconsole==5.4.2",
    "requests==2.31.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rsa==4.9",
    "scikit-learn==1.4.2",
    "scipy==1.13.0",
    "six==1.16.0",
    "sniffio==1.3.0",
    "soupsieve==2.4.1",
    "stack-data==0.6.3",
    "starlette==0.27.0",
    "terminado==0.17.1",
    "threadpoolctl==3.4.0",
    "tinycss2==1.2.1",
    "tomli==2.0.1",
    "tomlkit==0.12.1",
    "tornado==6.3.1",
    "traitlets==5.11.2",
    "typing_extensions==4.8.0",
    "tzdata==2023.3",
    "uri-template==1.2.0",
    "urllib3==2.0.7",
    "uvicorn==0.23.2",
    "uvloop==0.17.0",
    "watchfiles==0.20.0",
    "wcwidth==0.2.8",
    "webcolors==1.13",
    "webencodings==0.5.1",
    "websocket-client==1.5.1",
    "websockets==11.0.3",
    "widgetsnbextension==4.0.7" 
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.urls]
Homepage = "https://github.com/CommitMaster5500/CommonAssessmentTool"
Issues = "https://github.com/CommitMaster5500/CommonAssessmentTool/issues"

[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true